[refactor] Move ``_compare_eq_cls`` to its own module
